restore old migration
